# specs/001-purple-glow-social/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Purple Glow Social API
  version: 1.0.0
  description: |
    RESTful API for Purple Glow Social - AI-powered social media management platform.

    **Authentication**: All endpoints (except /auth/*) require Bearer token in Authorization header.

    **Rate Limiting**: 
    - Authenticated: 100 req/min per user
    - Unauthenticated: 10 req/min per IP

    **Error Responses**: All endpoints return RFC 7807 Problem Details for HTTP APIs format.
  contact:
    name: Purple Glow Social Support
    email: support@purpleglowsocial.com
  license:
    name: Proprietary
servers:
  - url: https://api.purpleglowsocial.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.purpleglowsocial.com/errors/validation-error'
        title:
          type: string
          example: 'Validation Error'
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: 'The request body contains invalid data.'
        instance:
          type: string
          format: uri
          example: '/api/v1/posts'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum: [user, business_admin, team_member]
        account_status:
          type: string
          enum: [active, suspended, deleted]
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
        notification_preferences:
          type: object
          properties:
            email_enabled:
              type: boolean
            in_app_enabled:
              type: boolean
            events:
              type: array
              items:
                type: string
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/unauthorized'
            title: 'Unauthorized'
            status: 401
            detail: 'Invalid or expired authentication token.'
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/forbidden'
            title: 'Forbidden'
            status: 403
            detail: 'You do not have permission to access this resource.'
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/not-found'
            title: 'Not Found'
            status: 404
            detail: 'The requested resource does not exist.'
...
security:
  - BearerAuth: []
paths:
...
  /analytics/summary:
    get:
      tags:
        - Analytics
      summary: Get analytics summary
      description: Get aggregated metrics for business profile
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: platform
          in: query
          schema:
            type: string
            enum: [facebook, instagram, twitter, linkedin]
      responses:
        '200':
          description: Summary retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_posts:
                    type: integer
                  published_posts:
                    type: integer
                  total_likes:
                    type: integer
                  total_comments:
                    type: integer
                  total_shares:
                    type: integer
                  avg_engagement_rate:
                    type: number
                  by_platform:
                    type: object
                    additionalProperties:
                      type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      description: Admin-only endpoint to view users
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, deleted]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/users/{id}/suspend:
    post:
      tags:
        - Admin
      summary: Suspend user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User suspended
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...